feat: acces token

+++ src/pages/api/[[...user]].tsx
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "GET") {
@@ -11,19 +12,35 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     res.status(200).json({ status: true, message: "Success", ststusCode: 200, data });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData("users", id, data, (status: boolean) => {
-      if (status) {
-        res.status(200).json({ status: true, message: "Success" });
+    const data = req.body;
+    const { user }: any = req.query;
+    const token: any = req.headers.authorization?.split(" ")[1];
+
+    jwt.verify(token, process.env.NEXTAUTH_SECREET || "", async (err: any, decoded: any) => {
+      if (decoded && decoded.role === "admin") {
+        await updateData("users", user[1], data, (status: boolean) => {
+          if (status) {
+            res.status(200).json({ status: true, message: "Success" });
+          } else {
+            res.status(400).json({ status: false, message: "Failed" });
+          }
+        });
       } else {
         res.status(400).json({ status: false, message: "Failed" });
       }
     });
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    await deleteData("users", user[1], (status: boolean) => {
-      if (status) {
-        res.status(200).json({ status: true, message: "Success" });
+    const token: any = req.headers.authorization?.split(" ")[1];
+    jwt.verify(token, process.env.NEXTAUTH_SECREET || "", async (err: any, decoded: any) => {
+      if (decoded && decoded.role === "admin") {
+        await deleteData("users", user[1], (status: boolean) => {
+          if (status) {
+            res.status(200).json({ status: true, message: "Success" });
+          } else {
+            res.status(400).json({ status: false, message: "Failed" });
+          }
+        });
       } else {
         res.status(400).json({ status: false, message: "Failed" });
       }
